Featured image changes
extra field: upload a featured image.
The featured image will be displayed in the widget (smaller) and on the
event list page.

the images are floating next to the date:

date - image - event details

They are floating like this on the event list page as well as in the
widget,

Depending on the design of your theme: they will have to have room to
float!

I was thinking of making a responsive version, but it all depends how
(responsive) the theme is set up