Update filterbar.php
Added switch option "months" to function "show" to add a dropdown with yyyy-mm strings as values to select single months from the filterbar. 
The dropdown is created by the function "show_months" as in function show_cats, just with yyyy-mm as values.

It can be used by setting "filterbar_items=months_dropdown" just like the category-dropdown.