adds an excerpt option
This adds an excerpt option to the addon.  This can be used in shortcode displays and the widget.

The excerpt is first generated from the custom excerpt field in the post.  If that is not there, it then checks the content for a read more tag and uses the content before the tag.  Finally if that is not there, it will use the main content and if you have entered a content_length number, it will truncate to that length.