Enable category filtering on ICS feeds (WIP)
For more adaptable usage of the ICS feed of events, the category filter
of shortcodes are passed to the instance of class 'EL_iCal'. This way,
any linked '.ics' file can contain only events from the given category.

For this to work, the shortcode's parameter 'cat_filter' is passed to
'EL_iCal::get_instance', where the category filter is prefixed to the
base name that is configured in the UI backend. Furthermore, there is
not only a single instance of 'EL_iCal' but an instance pool, organized
by the category slug (note: the slug for no filtering is 'all'). When
calling 'ical_init', all instances are created and initialized (i.e.,
'add_feed' is called appropriately).

Signed-off-by: Claus Hunsen <github@claushunsen.de>
Signed-off-by: Christian Hunsen <christian.hunsen@gmail.com>